Fix per-state holiday lists
I didn't notice that my own config was using a different version of
the code versus what I ultimately published. Because
`australia-holidays is a quoted list, we end up double-quoting here
when it gets interpolated into the per-state list by
calendar-holiday-list (specifically, it gets eval'd back to a quoted
list).