Add Friday before AFL Grand Final to VIC holidays
The AFL Grand Final is on the afternoon of the last Saturday of
September, with the Friday before it being a holiday in Victoria.
Using `(holiday-float 9 5 -1)` wouldn't work directly, as that Friday
may be the 30th; so, instead specify that it should be the last Friday
before the 29th (inclusive) and not the end of the month.

This should be good enough to work for the most part, but will not
handle cases where the AFL Grand Final is rescheduled (e.g. during
COVID, where it was on the 24th of October instead).

  (let ((calendar-holidays australia-holidays-for-vic))
    (list-holidays 2022))

  Friday, September 23, 2022: Friday before AFL Grand Final